ames_schools {AmesHousing}	R Documentation

Ames Public Schools

Description

Locations of local schools and outline of elementary school districts. Elementary school district boundaries are returned as sf object in CRS 4326

Details

The data set includes the school name and the geocodes.

Value

ames_schools_geo
a tibble

[Package AmesHousing version 0.0.4.9000 Index]